feat: add initial structure
Created initial structure of the site with basic features - optimization
of html and css, custom font, some markup and content. It'd be great to
avoid the middle step and pull css directly from `tailwind.css` file,
since it gets optimized on every change anyway. It takes 2s tops now, so
it's not an issue but may need to be changed in the future.